add info logs, deduplicate copy to static files part
